Models for Resume PDF generation
I'm going to make my own resume via PDF SHARP, first, I'll do the scheme here, and when I've cooked it, I'll use it on inventory-api. I'll need to make a table for all the data, that's going to be used for this pdf, afterwards, I'll add a button to generate the resume, and CRUD to modify everything I need on it.

Made ContactData.cs
Made Experience.cs
Made ResumeModel.cs